+++ src/all_Weapons/Bows/Bow.cpp
@@ -3,7 +3,7 @@
 #include "../../Entities/Enemy.hpp"
 #include <cmath>
 
-Bow::Bow(std::string name, std::string type, int dmg, float range, int as, int hands) : Weapon(name, type, dmg, range, as, hands) {}
+Bow::Bow(std::string name, std::string type, int dmg, float range, float as, int hands) : Weapon(name, type, dmg, range, as, hands) {}
 
 void Bow::attack(Enemy &target, Hero &hero) {
     sf::Vector2f heroPos = sf::Vector2f(hero.getX(), hero.getY());
