+++ src/all_Weapons/Bows/Bow.hpp
@@ -6,7 +6,7 @@ class Enemy;
 
 class Bow : public Weapon {
     public:
-        Bow(std::string name, std::string type, int dmg, float range, int as, int hands);
+        Bow(std::string name, std::string type, int dmg, float range, float as, int hands);
         void attack(Enemy &target, Hero &hero) override;
         ~Bow() override;
         Weapon* clone() const override;
